+++ aquadoggo/src/rpc/methods/entry_args.rs
@@ -22,7 +22,9 @@ pub async fn get_entry_args(
 
 #[cfg(test)]
 mod tests {
-    use crate::server::{build_server, ApiState};
+    use crate::config::Configuration;
+    use crate::context::Context;
+    use crate::server::build_server;
     use crate::test_helpers::{
         handle_http, initialize_db, random_entry_hash, rpc_error, rpc_request, rpc_response,
         TestClient,
@@ -33,8 +35,8 @@ mod tests {
     #[tokio::test]
     async fn respond_with_wrong_author_error() {
         let pool = initialize_db().await;
-        let state = ApiState::new(pool.clone());
-        let app = build_server(state);
+        let context = Context::new(pool, Configuration::default());
+        let app = build_server(context);
         let client = TestClient::new(app);
 
         let request = rpc_request(
@@ -55,8 +57,8 @@ mod tests {
     #[tokio::test]
     async fn get_entry_arguments() {
         let pool = initialize_db().await;
-        let state = ApiState::new(pool.clone());
-        let app = build_server(state);
+        let context = Context::new(pool, Configuration::default());
+        let app = build_server(context);
         let client = TestClient::new(app);
 
         let request = rpc_request(