+++ aquadoggo/src/runtime.rs
@@ -3,9 +3,11 @@
 use anyhow::Result;
 
 use crate::config::Configuration;
+use crate::context::Context;
 use crate::db::{connection_pool, create_database, run_pending_migrations, Pool};
-use crate::server::{start_server, ApiState};
-use crate::task::TaskManager;
+use crate::server::http_service;
+use crate::service_manager::ServiceManager;
+use crate::service_message::ServiceMessage;
 
 /// Makes sure database is created and migrated before returning connection pool.
 async fn initialize_db(config: &Configuration) -> Result<Pool> {
@@ -32,30 +34,34 @@ async fn initialize_db(config: &Configuration) -> Result<Pool> {
 #[allow(missing_debug_implementations)]
 pub struct Runtime {
     pool: Pool,
-    task_manager: TaskManager,
+    manager: ServiceManager<Context, ServiceMessage>,
 }
 
 impl Runtime {
     /// Start p2panda node with your configuration. This method can be used to run the node within
     /// other applications.
     pub async fn start(config: Configuration) -> Self {
-        let mut task_manager = TaskManager::new();
-
         // Initialize database and get connection pool
         let pool = initialize_db(&config)
             .await
             .expect("Could not initialize database");
 
-        // Initialize API state with shared connection pool
-        let api_state = ApiState::new(pool.clone());
+        // Create service manager with shared data between services
+        let context = Context::new(pool.clone(), config);
+        let mut manager = ServiceManager::<Context, ServiceMessage>::new(1024, context);
 
-        // Start JSON RPC API server
-        task_manager.spawn("API Server", async move {
-            start_server(&config, api_state).await?;
-            Ok(())
-        });
+        // Start HTTP server with GraphQL API
+        manager.add("http", http_service);
 
-        Self { pool, task_manager }
+        Self { pool, manager }
+    }
+
+    /// This future resolves when at least one system service stopped.
+    ///
+    /// It can be used to exit the application as a stopped service usually means that something
+    /// went wrong.
+    pub async fn on_exit(&self) {
+        self.manager.on_exit().await;
     }
 
     /// Close all running concurrent tasks and wait until they are fully shut down.
@@ -64,6 +70,6 @@ impl Runtime {
         self.pool.close().await;
 
         // Wait until all tasks are shut down
-        self.task_manager.shutdown().await;
+        self.manager.shutdown().await;
     }
 }